requirements.txt
asgiref==3.2.10
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.29
django-bootstrap3==14.1.0
django-crispy-forms==1.9.2
django-registration==2.4.1
django-tinymce==3.0.2
djangorestframework==3.11.1
gunicorn==20.0.4
importlib-metadata==1.7.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
zipp==3.1.0
